+++ ackpine-core/src/main/kotlin/ru/solrudev/ackpine/impl/uninstaller/activity/UninstallActivity.kt
@@ -21,7 +21,6 @@ import android.content.Intent
 import android.os.Bundle
 import androidx.annotation.RestrictTo
 import ru.solrudev.ackpine.impl.activity.SessionCommitActivity
-import ru.solrudev.ackpine.session.Session
 import ru.solrudev.ackpine.uninstaller.PackageUninstaller
 import ru.solrudev.ackpine.uninstaller.UninstallFailure
 
@@ -48,11 +47,7 @@ internal class UninstallActivity : SessionCommitActivity<UninstallFailure>(
 		ackpinePackageUninstaller = PackageUninstaller.getInstance(this)
 		super.onCreate(savedInstanceState)
 		if (packageNameToUninstall == null) {
-			withCompletableSession { session ->
-				session?.completeExceptionally(
-					IllegalStateException("$TAG: packageNameToUninstall was null.")
-				)
-			}
+			completeSessionExceptionally(IllegalStateException("$TAG: packageNameToUninstall was null."))
 			finish()
 			return
 		}
@@ -67,11 +62,8 @@ internal class UninstallActivity : SessionCommitActivity<UninstallFailure>(
 		if (requestCode != this.requestCode) {
 			return
 		}
-		val success = uninstallPackageContract.parseResult(this, resultCode)
-		val result = if (success) Session.State.Succeeded else Session.State.Failed(UninstallFailure.Generic)
-		withCompletableSession { session ->
-			session?.complete(result)
-		}
+		val result = uninstallPackageContract.parseResult(this, resultCode)
+		completeSession(result)
 	}
 
 	@SuppressLint("RestrictedApi")